simple_agent_01/pyproject.toml
[project]
name = "31-aug"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.10",
]
